init Todo components: list and its data
The very external component TodoApp contains TodoList and temporary Todo
datas as state. As props TodoApp pass in data to TodoList.

Inside TodoList, it runs a loop through the passed data then the loop
funtion pass in datas as a props to Todo.jsx.

Especialy when a loop goes through over a component, there has to be a
key value to be passed in to differentiate each instances. Also every
data can be passed with es6 spread syntax('...todo' in this case).