update check handling function:
Each Todo instance now have completed props. The completed prop
represents whether the checkbox is checked or not.

Once the Todo <div> is clicked, with onClick handler, it goes all the
way up to TodoApp and pass in id to handleToggle. Right here, it changes
the completed status of specific todo with the id and generates new Todo
arrays(updatedTodos).