Init TodoAPI: API to storage datas
API methods were created which handles fetching datas(getTodos) and
updating datas(setTodos).

In getTodos, it parses the original string data into object. On the
other hand setTodos stringify the object into strings to store in the
localStorage.

Each method is called inside the TodoApp.jsx whenever its needed.
getTodos only have to be called when browser is rendered first. After,
when user make a change on the state of datas, by the
componentDidUpdate lifecycle method, the setTodo() is called.